TOPU-6 feat: Rename package
