+++ src/main/java/jp/co/topucomunity/backend_java/users/domain/UserSession.java
@@ -1,0 +1,4 @@
+package jp.co.topucomunity.backend_java.users.domain;
+
+public record UserSession(String id) {
+}
