--- src/main/java/jp/co/topucomunity/backend_java/users/domains/UserSession.java
@@ -1,4 +1,0 @@
-package jp.co.topucomunity.backend_java.users.domains;
-
-public record UserSession(String id) {
-}
